% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakid.R
\name{detect_plumes}
\alias{detect_plumes}
\title{Detects plumes in a concentration time series.}
\usage{
detect_plumes(
  concentration,
  background,
  time,
  plume_sd_threshold = 3,
  plume_sd_starting = 2,
  plume_buffer = 10
)
}
\arguments{
\item{concentration}{Concentration time-series as a vector.}

\item{background}{The smoothed background time-series, as can be
obtained from \code{identify_background}. Must have the same
length as \code{concentration}.}

\item{time}{The time-series as a vector of POSIX or nanotime.
Must have the same length as \code{concentration}.}

\item{plume_sd_threshold}{The number of standard deviations that a sample
must exceed to be defined as a plume.}

\item{plume_sd_starting}{Once a plume has been identified due to it
crossing \code{plume_sd_threshold}, its duration is considered as the
times when it is more than this many standard deviatiations above the
background.}

\item{plume_buffer}{A buffer in seconds applied to plumes, so
that if they are overlapping they are merged into the same plume.}
}
\value{
A Data Frame where each row corresponds to a unique plume, whose
time boundaries are contained in the 2 columns: \code{start} and \code{end}.
}
\description{
Detects plumes in a concentration time series.
}
