% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakid.R
\name{identify_background}
\alias{identify_background}
\title{Identifies background from a concentration time-series.}
\usage{
identify_background(
  concentration,
  bg_sd_window = 180,
  bg_sd_threshold = 0.5,
  bg_mean_window = 660
)
}
\arguments{
\item{concentration}{Concentration time-series as a vector.}

\item{bg_sd_window}{Window size for the rolling standard deviation
smooth to identify the background, as an integer.}

\item{bg_sd_threshold}{Background measurements are considered as
those whose rolling sd is within this threshold}

\item{bg_mean_window}{The rolling mean to smooth the interpolated
background, as an integer.}
}
\value{
A vector with the same length as \code{concentration} containing the
smoothed background.
}
\description{
This process takes 3 steps:
\enumerate{
\item Obtaining a smoothed rolling standard deviation of the background
\item Identifying background measurements as those that lie within a set
threshold
\item Interpolating values outside of this background threshold in a
linear manner and then using a rolling mean so that 'background'
measurements are available for the entire time-series.
}
}
