% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakid.R
\name{plot_background}
\alias{plot_background}
\title{Diagnostic plot to aid background extraction}
\usage{
plot_background(
  concentration,
  time,
  background,
  plume_sd_threshold = 3,
  plume_sd_starting = 2,
  ylabel = "Concentration",
  xlabel = "Time (UTC)",
  date_fmt = "\%H:\%M",
  bg_alpha = 0.5
)
}
\arguments{
\item{concentration}{Concentration time-series as a vector.}

\item{time}{The time-series as a vector of POSIX or nanotime.
Must have the same length as \code{concentration}.}

\item{background}{The smoothed background time-series object, as can be
obtained from \code{identify_background}}

\item{plume_sd_threshold}{The number of standard deviations that a sample
must exceed to be defined as a plume.}

\item{plume_sd_starting}{Once a plume has been identified due to it
crossing \code{plume_sd_threshold}, its duration is considered as the
times when it is more than this many standard deviations above the
background.}

\item{ylabel}{y-axis label}

\item{xlabel}{x-axis label}

\item{date_fmt}{Format of the x-axis labels, see \code{date_labels}
argument of \code{ggplot2::scale_x_datetime}.}

\item{bg_alpha}{The alpha level of the background concentration.}
}
\value{
A \code{ggplot2} object, so it can be modified further.
}
\description{
Plots the concentration time-series highlighting the extracted background
(red), the threshold for what is considered a plume (orange), and when
plumes will be determined to have started at (blue).
}
