% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakid.R
\name{plot_plumes}
\alias{plot_plumes}
\title{Plots plumes against the background concentration.}
\usage{
plot_plumes(
  conc,
  times,
  plumes,
  ylabel = "Concentration",
  xlabel = "Time (UTC)",
  date_fmt = "\%H:\%M",
  bg_alpha = 0.5
)
}
\arguments{
\item{conc}{The concentration time-series as a vector}

\item{times}{The sample times corresponding to the concentrations, must have
the same length as \code{conc}.}

\item{plumes}{A Data Frame with 'start' and 'end' columns
containing plume boundaries, as returned by \code{detect_plumes}}

\item{ylabel}{y-axis label}

\item{xlabel}{x-axis label}

\item{date_fmt}{Format of the x-axis labels, see \code{date_labels}
argument of \code{ggplot2::scale_x_datetime}.}

\item{bg_alpha}{The alpha level of the background concentration.}
}
\value{
A \code{ggplot2} object, so it can be modified further.
}
\description{
Plots plumes against the background concentration.
}
